First step of cleanups for GeoRef initializations
The default Georeferences should be obtained only once,
during the initialization, and only if the GeoRef property
did not already have a value.

For now, this is done in PostInitProperties - BUT then,
it will not be possible to save objects with a 'None'
GeoRef - this has to be reviewed.

When the GeoRef is changed in the UI, the GeoRefUpdated
delegates should be added.